refactor: add dtos to users request operations
