+++ src/auth/strategies/local.strategy.ts
@@ -2,7 +2,7 @@ import { Strategy } from 'passport-local';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { AuthService } from '../auth.service';
-import { UserDto } from 'src/user/dto/user-response';
+import { UserDto } from '../../user/dto/user-response';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
