+++ src/user/dto/create-user-request.ts
@@ -1,0 +1,4 @@
+export class CreateUserRequest {
+  email: string;
+  password: string;
+}
