+++ src/user/dto/update-user-request.ts
@@ -1,0 +1,12 @@
+import { Role } from '@prisma/client';
+
+export interface UpdateUserRequest {
+  where: {
+    id: number;
+  };
+  data: {
+    email?: string;
+    password?: string;
+    role?: Role;
+  };
+}
