+++ src/user/user.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { Prisma } from '@prisma/client';
-import { UserDto } from './dto/user-response';
+import { UserDto } from './dto';
 
 @Injectable()
 export class UserRepository {
