+++ src/user/user.service.ts
@@ -1,24 +1,21 @@
 import { Injectable } from '@nestjs/common';
-import { Prisma } from '@prisma/client';
 import { UserRepository } from './user.repository';
 import { UserDto } from './dto/user-response';
+import { UpdateUserRequest, CreateUserRequest } from './dto';
 
 @Injectable()
 export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
 
-  createUser(data: Prisma.UserCreateInput): Promise<UserDto> {
+  createUser(data: CreateUserRequest): Promise<UserDto> {
     return this.userRepository.createUser(data);
   }
 
-  updateUser(
-    where: Prisma.UserWhereUniqueInput,
-    data: Prisma.UserUpdateInput,
-  ): Promise<UserDto> {
-    return this.userRepository.updateUser(where, data);
+  updateUser(payload: UpdateUserRequest): Promise<UserDto> {
+    return this.userRepository.updateUser(payload.where, payload.data);
   }
 
-  deleteUser(where: Prisma.UserWhereUniqueInput): Promise<void> {
-    return this.userRepository.removeUserById(where.id);
+  deleteUser(id: number): Promise<void> {
+    return this.userRepository.removeUserById(id);
   }
 }
